Warning cleanup
git-svn-id: https://svn.outpostuniverse.org:8443/svn/outposthd@269 bdf9137d-f3be-7946-88fe-7b0bbfdbf31a

+++ src/UI/Core/TextField.cpp
@@ -168,8 +168,8 @@ void TextField::drawCursor()
 		{
 			// updateCursor() should be called only on events relating to the cursor so this is temporary.
 			updateCursor();
-			Utility<Renderer>::get().drawLine(mCursorX + 1, rect().y() + FIELD_PADDING + 1, mCursorX + 1, rect().y() + rect().height() - FIELD_PADDING, 0, 0, 0);
-			Utility<Renderer>::get().drawLine(mCursorX, rect().y() + FIELD_PADDING, mCursorX, rect().y() + rect().height() - FIELD_PADDING - 1, 255, 255, 255);
+			Utility<Renderer>::get().drawLine(static_cast<float>(mCursorX + 1), rect().y() + FIELD_PADDING + 1, static_cast<float>(mCursorX + 1), rect().y() + rect().height() - FIELD_PADDING, 0, 0, 0);
+			Utility<Renderer>::get().drawLine(static_cast<float>(mCursorX), rect().y() + FIELD_PADDING, static_cast<float>(mCursorX), rect().y() + rect().height() - FIELD_PADDING - 1, 255, 255, 255);
 		}
 		
 		if(mCursorTimer.accumulator() > CURSOR_BLINK_DELAY)
@@ -186,7 +186,7 @@ void TextField::drawCursor()
  */
 void TextField::drawTextHighlight()
 {
-	Utility<Renderer>::get().drawBoxFilled(rect().x() + FIELD_PADDING, rect().y(), font().width(text()), rect().height(), 0, 0, 150, 100);
+	Utility<Renderer>::get().drawBoxFilled(rect().x() + FIELD_PADDING, rect().y(), static_cast<float>(font().width(text())), rect().height(), 0, 0, 150, 100);
 }
 
 
@@ -203,7 +203,7 @@ void TextField::updateCursor()
 		mScrollOffset = 0;
 
 
-	mCursorX = rect().x() + FIELD_PADDING + cursorX - mScrollOffset;
+	mCursorX = static_cast<int>(rect().x() + FIELD_PADDING + cursorX - mScrollOffset);
 }
 
 
@@ -326,7 +326,7 @@ void TextField::onMouseDown(EventHandler::MouseButton button, int x, int y)
 	else
 		hasFocus(true);
 
-	int relativePosition = x - rect().x();
+	int relativePosition = static_cast<int>(x - rect().x());
 
 	// If the click occured past the width of the text, we can immediatly
 	// set the position to the end and move on.
